# swagger.yaml
openapi: 3.0.0
info:
  title: DiscoverIT API
  description: Node.js + MongoDB - DiscoverIT API with accounts and nagios monitor
  version: 1.0.0
servers:
  - url: http://localhost:70
    description: Local development server
  - url: http://169.59.166.94:70
    description: cloud development server
paths:
...
  /inventario:
    get:
      summary: get array of equipment in inventory
      description: users can acces their inventory
      operationId: getEquipment
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Details of the specified account
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: "string"
                    example: "xxxxxxxxxxxxxxxxx"
                  Serial:
                    type: string
                    example: "xxxxxxxx"
                  Brand:
                    type: string
                    example: "Alpha"
                  City:
                    type: string
                    example: "street 000, CDMX, 01210"
                  Contrato:
                    type: string
                    example: "xxxxxxx"
                  Equipment:
                    type: string
                    example: "Server"
                  Model:
                    type: string
                    example: "5000x"
                  SLA:
                    type: string
                    example: "24x7xNBD"
                  ServiceTag:
                    type: string
                    example: "nan"
                  Start:
                    type: string
                    example: "12-05-21"
                  end:
                    type: string
                    example: "12-05-22"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      summary: Create a new contract and equipment
      description: Restricted to admin users.
      operationId: createEquipment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Serial:
                    type: string
                    example: "xxxxxxxx"
                Brand:
                  type: string
                  example: "Alpha"
                City:
                  type: string
                  example: "street 000, CDMX, 01210"
                Contrato:
                  type: string
                  example: "xxxxxxx"
                Equipment:
                  type: string
                  example: "Server"
                Model:
                  type: string
                  example: "5000x"
                SLA:
                    type: string
                    example: "24x7xNBD"
                ServiceTag:
                    type: string
                    example: "nan"
                Start:
                  type: string
                  example: "12-05-21"
                End:
                  type: string
                  example: "12-05-22"
              required:
                - Serial
                - Brand
                - City
                - Contrato
                - Equipment
                - Model
                - SLA
                - ServiceTag
                - Start
                - End
      responses:
        "200":
          description: Contract and equipment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  Serial:
                    type: string
                    example: "xxxxxxxx"
                  Brand:
                    type: string
                    example: "Alpha"
                  City:
                    type: string
                    example: "street 000, CDMX, 01210"
                  Contrato:
                    type: string
                    example: "xxxxxxx"
                  Equipment:
                    type: string
                    example: "Server"
                  Model:
                    type: string
                    example: "5000x"
                  SLA:
                      type: string
                      example: "24x7xNBD"
                  ServiceTag:
                      type: string
                      example: "nan"
                  Start:
                    type: string
                    example: "12-05-21"
                  End:
                    type: string
                    example: "12-05-22"
        "400":
          description: Serial number is already registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Serial number is already registered"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"